pipeline/requirements.txt
pandas
numpy
thefuzz
python-levenshtein
psycopg2-binary
python-dotenv
requests
